+++ app/src/main/java/com/vk/vsvans/BlogShop/model/fns/Receipt.kt
@@ -25,7 +25,7 @@ data class Receipt(
     val retailPlaceAddress: String,
     val shiftNumber: Int,
     val taxationType: Int,
-    val totalSum: Int,
+    val totalSum: Double,
     val user: String,
     val userInn: String
 )
